Create template controller to handle all htmls
