+++ src/main/java/com/example/tp_dos_desi/model/Ciudad.java
@@ -1,13 +1,10 @@
 package com.example.tp_dos_desi.model;
 
-import java.util.List;
-
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.OneToMany;
 
 @Entity
 public class Ciudad {
