+++ src/main/java/com/example/tp_dos_desi/repository/AvionRepository.java
@@ -1,12 +1,9 @@
 package com.example.tp_dos_desi.repository;
 
-import com.example.tp_dos_desi.model.Asiento;
 import com.example.tp_dos_desi.model.Avion;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
 
 @Repository
 public interface AvionRepository extends JpaRepository<Avion, Long> {
