+++ src/main/java/com/example/tp_dos_desi/repository/VueloRepository.java
@@ -2,11 +2,7 @@ package com.example.tp_dos_desi.repository;
 
 import com.example.tp_dos_desi.model.Vuelo;
 
-import java.time.LocalDate;
-import java.util.List;
-
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
