+++ src/main/java/com/example/tp_dos_desi/service/AsientoService.java
@@ -1,12 +1,9 @@
 package com.example.tp_dos_desi.service;
 
 import com.example.tp_dos_desi.model.Asiento;
-import com.example.tp_dos_desi.model.Vuelo;
 import com.example.tp_dos_desi.repository.AsientoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 public class AsientoService {
@@ -26,7 +23,7 @@ public class AsientoService {
         return this.asientoRepository.save(asiento);
     }
 
-    public Asiento buscarPorId(Long id){
+    public Asiento buscarPorId(Long id) {
         return this.asientoRepository.getById(id);
     }
