+++ src/main/resources/templates/asientoComprado.html
@@ -2,42 +2,42 @@
 <html xmlns:th="http://www.thymeleaf.org">
 
 <head>
-  <meta charset="UTF-8">
-  <title>Lista de vuelos</title>
-  <script th:src="@{/webjars/jquery/jquery.min.js}"></script> <!-- Include jQuery -->
+    <meta charset="UTF-8">
+    <title>Lista de vuelos</title>
+    <script th:src="@{/webjars/jquery/jquery.min.js}"></script> <!-- Include jQuery -->
 </head>
 
 <body>
 
-  <h1 th:text="'Vuelo ' + ${vuelo.numeroVuelo}"></h1>
-  <h2 th:text="${origen} + ' - ' + ${destino}"></h2>
-  <h3 th:text="'Fecha de salida: ' + ${vuelo.fechaHoraSalida}"></h3>
+    <h1 th:text="'Vuelo ' + ${vuelo.numeroVuelo}"></h1>
+    <h2 th:text="${origen} + ' - ' + ${destino}"></h2>
+    <h3 th:text="'Fecha de salida: ' + ${vuelo.fechaHoraSalida}"></h3>
 
-  <h1>Asiento comprado</h1>
-  <h2 th:text="'Numero: '+${asiento.numeroAsiento}"></h2>
-  <h2 th:text="'Precio: '+${asiento.precio}"></h2>
+    <h1>Asiento comprado</h1>
+    <h2 th:text="'Numero: '+${asiento.numeroAsiento}"></h2>
+    <h2 th:text="'Precio: '+${asiento.precio}"></h2>
 
-  <h1>Datos del cliente</h1>
-  <table>
-      <tr>
-          <td>Nombre:</td>
-          <td th:text="${cliente.nombre}"></td>
-      </tr>
-      <tr>
-          <td>Apellido:</td>
-          <td th:text="${cliente.apellido}"></td>
-      </tr>
-      <tr>
-          <td>Domicilio:</td>
-          <td th:text="${cliente.domicilio}"></td>
-      </tr>
-      <tr>
-          <td>E-mail:</td>
-          <td th:text="${cliente.email}"></td>
-      </tr>
-  </table>
+    <h1>Datos del cliente</h1>
+    <table>
+        <tr>
+            <td>Nombre:</td>
+            <td th:text="${cliente.nombre}"></td>
+        </tr>
+        <tr>
+            <td>Apellido:</td>
+            <td th:text="${cliente.apellido}"></td>
+        </tr>
+        <tr>
+            <td>Domicilio:</td>
+            <td th:text="${cliente.domicilio}"></td>
+        </tr>
+        <tr>
+            <td>E-mail:</td>
+            <td th:text="${cliente.email}"></td>
+        </tr>
+    </table>
 
-
+    <a th:href="'/vuelo/'+${vuelo.id}+'?dni='+${cliente.id}">Volver</a>
 </body>
 
 </html>
