+++ src/main/resources/templates/clienteEleccionVuelo.html
@@ -72,7 +72,7 @@
     </form>
 
 
-    </script>
+    <a href="/clientes"> Volver</a>
 </body>
 
 </html>
